Rename cloud resolver
