+++ sematic/function.py
@@ -306,7 +306,7 @@ def func(
     func: Optional[Callable]
         The `Callable` to instrument; usually the decorated function.
     standalone: bool
-        When using the `CloudResolver`, whether the instrumented function should
+        When using the `CloudRunner`, whether the instrumented function should
         be executed in a standalone Kubernetes Job or inside the same process
         and worker that is executing the `Resolver` itself.
 
@@ -319,7 +319,7 @@ def func(
 
         Do not activate this on a non-deterministic function!
     resource_requirements: Optional[ResourceRequirements]
-        When using the `CloudResolver`, specifies what special execution
+        When using the `CloudRunner`, specifies what special execution
         resources the function requires. Defaults to `None`.
     retry: Optional[RetrySettings]
         Specifies in case of which Exceptions the function's execution should be
